chore: Move action builder components.
