+++ src/components/Guilds/ActionsBuilder/common/ProposalOptionTag/ProposalOptionTag.test.tsx
@@ -1,4 +1,4 @@
-import { render } from '../../../../../../utils/tests';
+import { render } from '../../../../../utils/tests';
 import { ProposalOptionTag } from './ProposalOptionTag';
 
 describe('ProposalOptionTag', () => {
